requirements.txt
absl-py==2.2.2
accelerate==0.25.0
addict==2.4.0
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
av==14.3.0
basicsr==1.4.2
bitsandbytes==0.39.0
click==8.2.0
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.25.0
einops==0.7.0
exceptiongroup==1.3.0
fastapi==0.112.4
ffmpy==0.5.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
future==1.0.0
fvcore==0.1.5.post20221221
gradio==4.24.0
gradio_client==0.14.0
grpcio==1.71.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.25.1
humanfriendly==10.0
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iopath==0.1.10
Jinja2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
lmdb==1.6.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mkl-fft
mkl-service==2.4.1
narwhals==1.38.2
networkx
numpy
omegaconf==2.3.0
onnxruntime==1.16.2
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.8
portalocker==3.1.1
protobuf==6.30.2
psutil==7.0.0
pycocotools==2.0.8
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.9
safetensors==0.5.3
scikit-image==0.24.0
scipy==1.11.1
semantic-version==2.10.0
shellingham==1.5.4
six
sniffio==1.3.1
starlette==0.38.6
sympy
tabulate==0.9.0
tb-nightly==2.20.0a20250510
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.5.10
tokenizers==0.15.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==1.2.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.36.2
triton==2.0.0
typer==0.15.3
typing-inspection==0.4.0
tzdata==2025.2
urllib3
uvicorn==0.34.2
websockets==11.0.3
Werkzeug==3.1.3
yacs==0.1.8
yapf==0.43.0
zipp==3.21.0
